lambda/segment-sam/requirements.txt
numpy
opencv-python
Pillow==9.5.0
